Copied over some tests from TestConnector
Summary:
* Added TestSaveToMemSQLNoNodesAvailable
* Added SparkSQLTypesSpec
* Added ExtraColumnsSpec
* Added to SaveToMemSQLSpec
* Added to MemSQLTableSpec
* Changed the type conversion scheme from ShortType <-> TINYINT to ShortType <-> SMALLINT

Fixes T18834

Test Plan: Local

Reviewers: hurshal, allan, wsong, ankur, carl

Reviewed By: carl

Subscribers: engineering-list

Maniphest Tasks: T18834

Differential Revision: http://grizzly.memsql.com/D13592

+++ tests/src/main/scala/com/memsql/spark/TestTypes.scala
@@ -26,7 +26,8 @@ object TestTypes {
 
 class TestTypes extends TestApp with Logging {
   def types: Seq[TypeMapping] = Seq(
-    TypeMapping(ShortType, MemSQLType("TINYINT"), Seq(0.toShort, 16.toShort)),
+    TypeMapping(ByteType, MemSQLType("TINYINT"), Seq(0.toShort, 16.toShort)),
+
     TypeMapping(ShortType, MemSQLType("SMALLINT"), Seq(0.toShort, 64.toShort)),
 
     TypeMapping(IntegerType, MemSQLType("MEDIUMINT"), Seq(0, 256)),
